# contracts/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Handdraw API",
    "version": "1.0.0",
    "description": "Handdraw API contract. Identity, workspace onboarding and project/board routes are wired behind explicit configuration. Board deletion additionally requires the cleanup worker. Team membership, invitations and board Viewer grants require the membership feature flag. Invitation email delivery remains pending integration. Bearer identity is verified before profile resolution. Unknown fields/trailing JSON are rejected. Future domain endpoints are added with their implementation contracts."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8081",
      "description": "Local development; deployment must use HTTPS."
    }
  ],
  "security": [
    {
      "supabaseBearer": []
    }
  ],
...
  "paths": {
...
    "/v1/workspaces": {
      "get": {
        "operationId": "listWorkspaces",
        "summary": "List workspaces",
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Cursor"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "private, no-store"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "503": {
            "$ref": "#/components/responses/503"
          }
        },
        "x-max-request-bytes": 65536,
        "tags": [
          "Workspaces"
        ],
        "x-implementation-status": "available-when-enabled",
        "description": "Available when the corresponding runtime feature is enabled."
      },
      "post": {
        "operationId": "createWorkspace",
        "summary": "Create workspace",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "private, no-store"
                  ]
                }
              },
              "ETag": {
                "schema": {
                  "type": "string",
                  "pattern": "^\"[1-9][0-9]*\"$"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "503": {
            "$ref": "#/components/responses/503"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "413": {
            "$ref": "#/components/responses/413"
          },
          "415": {
            "$ref": "#/components/responses/415"
          },
          "422": {
            "$ref": "#/components/responses/422"
          },
          "202": {
            "description": "Operation already processing; result contains a safe operation reference. No replay of cached secrets.",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "private, no-store"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingOperationResponse"
                }
              }
            }
          }
        },
        "x-max-request-bytes": 65536,
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspace"
              }
            }
          }
        },
        "tags": [
          "Workspaces"
        ],
        "x-implementation-status": "available-when-enabled",
        "description": "Available when the corresponding runtime feature is enabled. Atomically create workspace + Owner; subscription is separate and starts unavailable."
      }
    },
    "/v1/workspaces/{workspace_id}": {
      "get": {
        "operationId": "getWorkspace",
        "summary": "Get workspace",
        "parameters": [
          {
            "$ref": "#/components/parameters/workspace_id"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "private, no-store"
                  ]
                }
              },
              "ETag": {
                "schema": {
                  "type": "string",
                  "pattern": "^\"[1-9][0-9]*\"$"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "503": {
            "$ref": "#/components/responses/503"
          }
        },
        "x-max-request-bytes": 65536,
        "tags": [
          "Workspaces"
        ],
        "x-implementation-status": "available-when-enabled",
        "description": "Available when the corresponding runtime feature is enabled."
      },
      "patch": {
        "operationId": "renameWorkspace",
        "summary": "Rename workspace",
        "parameters": [
          {
            "$ref": "#/components/parameters/workspace_id"
          },
          {
            "$ref": "#/components/parameters/IfMatch"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "enum": [
                    "private, no-store"
                  ]
                }
              },
              "ETag": {
                "schema": {
                  "type": "string",
                  "pattern": "^\"[1-9][0-9]*\"$"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "429": {
            "$ref": "#/components/responses/429"
          },
          "503": {
            "$ref": "#/components/responses/503"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "413": {
            "$ref": "#/components/responses/413"
          },
          "415": {
            "$ref": "#/components/responses/415"
          },
          "422": {
            "$ref": "#/components/responses/422"
          },
          "412": {
            "$ref": "#/components/responses/412"
          },
          "428": {
            "$ref": "#/components/responses/428"
          }
        },
        "x-max-request-bytes": 65536,
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Rename"
              }
            }
          }
        },
        "tags": [
          "Workspaces"
        ],
        "x-implementation-status": "available-when-enabled",
        "description": "Available when the corresponding runtime feature is enabled."
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "supabaseBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Supabase JWT"
      }
    },
    "schemas": {
...
      "WorkspaceID": {
        "type": "string",
        "pattern": "^ws_[0-9A-Za-z]{27}$",
        "format": "handdraw-resource-id",
        "description": "Canonical nonzero 160-bit KSUID; case sensitive. Regex is supplemented by the registered format validator."
      },
...
      "Revision": {
        "type": "string",
        "pattern": "^[1-9][0-9]*$",
        "format": "handdraw-revision",
        "description": "Positive signed int64 serialized as decimal text to retain JavaScript precision."
      },
      "Name": {
        "type": "string",
        "minLength": 1,
        "maxLength": 200,
        "pattern": "^\\S(?:[\\s\\S]*\\S)?$"
      },
      "Role": {
        "type": "string",
        "enum": [
          "owner",
          "editor",
          "viewer"
        ]
      },
...
      "Workspace": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "$ref": "#/components/schemas/WorkspaceID"
          },
          "name": {
            "$ref": "#/components/schemas/Name"
          },
          "kind": {
            "type": "string",
            "enum": [
              "personal",
              "team"
            ]
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "revision": {
            "$ref": "#/components/schemas/Revision"
          },
          "access_revision": {
            "$ref": "#/components/schemas/Revision"
          },
          "lifecycle": {
            "type": "string",
            "enum": [
              "ready",
              "purging",
              "deleted"
            ]
          },
          "capabilities": {
            "$ref": "#/components/schemas/Capabilities"
          },
          "entitlement": {
            "$ref": "#/components/schemas/Entitlement"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "kind",
          "role",
          "revision",
          "access_revision",
          "lifecycle",
          "capabilities",
          "entitlement",
          "created_at",
          "updated_at"
        ]
      },
...
      "CreateWorkspace": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": {
            "$ref": "#/components/schemas/Name"
          },
          "kind": {
            "type": "string",
            "enum": [
              "personal",
              "team"
            ]
          }
        },
        "required": [
          "name",
          "kind"
        ]
      },
...
      "Rename": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": {
            "$ref": "#/components/schemas/Name"
          }
        },
        "required": [
          "name"
        ]
      },
...
      "PendingOperation": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "operation": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "status": {
            "type": "string",
            "enum": [
              "processing"
            ]
          }
        },
        "required": [
          "operation",
          "status"
        ]
      },
      "Meta": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "request_id": {
            "type": "string",
            "minLength": 1
          },
          "pagination": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "next_cursor": {
                "type": "string",
                "maxLength": 4096,
                "nullable": true
              }
            },
            "required": [
              "next_cursor"
            ]
          }
        },
        "required": [
          "request_id"
        ]
      },
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "error": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "code": {
                "type": "string",
                "minLength": 1
              },
              "message": {
                "type": "string",
                "minLength": 1
              },
              "details": {
                "type": "object",
                "additionalProperties": true
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "success",
          "meta",
          "error"
        ]
      },
...
      "WorkspaceResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "result": {
            "$ref": "#/components/schemas/Workspace"
          }
        },
        "required": [
          "success",
          "meta",
          "result"
        ]
      },
...
      "PendingOperationResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          },
          "result": {
            "$ref": "#/components/schemas/PendingOperation"
          }
        },
        "required": [
          "success",
          "meta",
          "result"
        ]
      },
      "WorkspaceListResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "meta": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "request_id": {
                "type": "string",
                "minLength": 1
              },
              "pagination": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "next_cursor": {
                    "type": "string",
                    "maxLength": 4096,
                    "nullable": true
                  }
                },
                "required": [
                  "next_cursor"
                ]
              }
            },
            "required": [
              "request_id",
              "pagination"
            ]
          },
          "result": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Workspace"
            }
          }
        },
        "required": [
          "success",
          "meta",
          "result"
        ]
      },
...
    },
    "parameters": {
      "Limit": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 50
        }
      },
      "Cursor": {
        "name": "cursor",
        "in": "query",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 4096
        },
        "description": "Opaque authenticated cursor bound to actor, workspace, resource, filter and the operation-specific ordering described below; mismatch is invalid_cursor."
      },
      "IfMatch": {
        "name": "If-Match",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^\"[1-9][0-9]*\"$",
          "format": "handdraw-etag"
        },
        "description": "Strong quoted decimal metadata revision. Missing: 428. Stale: 412. Never the CRDT revision."
      },
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "format": "handdraw-idempotency-key"
        },
        "description": "Nonzero client UUID, scoped to actor+operation. Same normalized payload retries reuse the result after current authorization. Changed payload: 409; in progress: 202; 24-hour record TTL."
      },
      "workspace_id": {
        "name": "workspace_id",
        "in": "path",
        "required": true,
        "schema": {
          "$ref": "#/components/schemas/WorkspaceID"
        }
      },
...
    },
    "responses": {
      "400": {
        "description": "invalid_request / invalid_cursor / invalid_resource_id",
        "headers": {
          "X-Request-ID": {
            "schema": {
              "type": "string"
            }
          },
          "Cache-Control": {
            "schema": {
              "type": "string",
              "enum": [
                "private, no-store"
              ]
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "401": {
        "description": "unauthenticated",
        "headers": {
          "X-Request-ID": {
            "schema": {
              "type": "string"
            }
          },
          "Cache-Control": {
            "schema": {
              "type": "string",
              "enum": [
                "private, no-store"
              ]
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "403": {
        "description": "permission_denied / entitlement_read_only",
        "headers": {
          "X-Request-ID": {
            "schema": {
              "type": "string"
            }
          },
          "Cache-Control": {
            "schema": {
              "type": "string",
              "enum": [
                "private, no-store"
              ]
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "404": {
        "description": "not_found (including inaccessible IDs)",
        "headers": {
          "X-Request-ID": {
            "schema": {
              "type": "string"
            }
          },
          "Cache-Control": {
            "schema": {
              "type": "string",
              "enum": [
                "private, no-store"
              ]
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "409": {
        "description": "idempotency_conflict / project_not_empty / board_initializing / membership_conflict",
        "headers": {
          "X-Request-ID": {
            "schema": {
              "type": "string"
            }
          },
          "Cache-Control": {
            "schema": {
              "type": "string",
              "enum": [
                "private, no-store"
              ]
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "412": {
        "description": "revision_conflict",
        "headers": {
          "X-Request-ID": {
            "schema": {
              "type": "string"
            }
          },
          "Cache-Control": {
            "schema": {
              "type": "string",
              "enum": [
                "private, no-store"
              ]
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "413": {
        "description": "payload_too_large",
        "headers": {
          "X-Request-ID": {
            "schema": {
              "type": "string"
            }
          },
          "Cache-Control": {
            "schema": {
              "type": "string",
              "enum": [
                "private, no-store"
              ]
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "415": {
        "description": "unsupported_media_type",
        "headers": {
          "X-Request-ID": {
            "schema": {
              "type": "string"
            }
          },
          "Cache-Control": {
            "schema": {
              "type": "string",
              "enum": [
                "private, no-store"
              ]
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "422": {
        "description": "invalid_document_schema",
        "headers": {
          "X-Request-ID": {
            "schema": {
              "type": "string"
            }
          },
          "Cache-Control": {
            "schema": {
              "type": "string",
              "enum": [
                "private, no-store"
              ]
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "428": {
        "description": "precondition_required",
        "headers": {
          "X-Request-ID": {
            "schema": {
              "type": "string"
            }
          },
          "Cache-Control": {
            "schema": {
              "type": "string",
              "enum": [
                "private, no-store"
              ]
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "429": {
        "description": "rate_limited",
        "headers": {
          "X-Request-ID": {
            "schema": {
              "type": "string"
            }
          },
          "Cache-Control": {
            "schema": {
              "type": "string",
              "enum": [
                "private, no-store"
              ]
            }
          },
          "Retry-After": {
            "schema": {
              "type": "string",
              "pattern": "^[0-9]+$"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "503": {
        "description": "identity_unavailable / dependency_unavailable / save_unavailable",
        "headers": {
          "X-Request-ID": {
            "schema": {
              "type": "string"
            }
          },
          "Cache-Control": {
            "schema": {
              "type": "string",
              "enum": [
                "private, no-store"
              ]
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  }
}